move JSON SA types to deprecated namespace
