+++ db/functions/known_db_functions.py
@@ -13,8 +13,6 @@ from db.deprecated.utils import get_module_members_that_satisfy
 import db.functions.base
 import db.functions.packed
 import db.deprecated.types.custom
-import db.types.custom.json_array
-import db.types.custom.json_object
 
 from db.functions.base import DBFunction
 
@@ -31,8 +29,6 @@ _modules_to_search_in = tuple([
     db.deprecated.types.custom,
     db.functions.base,
     db.functions.packed,
-    db.types.custom.json_array,
-    db.types.custom.json_object,
 ])
